requirements.txt
aiohttp==3.9.3
beautifulsoup4==4.12.3
bs4==0.0.2
Django==5.0.2
django-jazzmin==2.6.0
django-solo==2.2.0
Pillow
